build: allow concurrent runs
This allows running builds in parallel. The basic change that enables
this is:

    runPolicy: Parallel

The rest of the code change here is to deal with it in the main build
job in a smart-ish way:
1. we use a lock to make sure only one build can run at a time per
   stream; we don't want multiple runs clobbering each others' results
2. we use `inversePrecedence: true` so that if multiple builds for a
   stream are waiting (this should rarely happen, but could if e.g. a
   build gets stuck for some reason), the most recent build started
   takes the lock first, and all the other builds are automatically
   skipped

Related to 2., I'm increasingly leaning towards moving away from
buildconfigs and just have a job DSL define a separate job per stream.
The CI rework I'm working on will leverage this too. This will fix a few
things, including:
- dropping the "meta" wrapper jobs
  (Jenkinsfile.{production,mechanical,development})
- fixing bugs related to the OpenShift Jenkins plugin (e.g. how we have
  to trigger other jobs using `oc start-build` instead of the `build`
  step)
- making it trivial to see logs from the last e.g. stable/testing build
  and not worrying about having them be GC'ed
- and finally, we can then drop the logic from this patch since it's
  obviously correct